% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counting.R
\name{df_counting}
\alias{df_counting}
\title{Creates counting process dataset}
\usage{
df_counting(
  df,
  tte.name,
  event.name,
  treat.name,
  weight.name = NULL,
  strata.name = NULL,
  arms = c("treat", "control"),
  time.zero = 0,
  tpoints.add = c(0),
  by.risk = 6,
  time.zero.label = 0,
  risk.add = NULL,
  get.cox = TRUE,
  cox.digits = 2,
  lr.digits = 2,
  cox.eps = 0.001,
  lr.eps = 0.001,
  qprob = 0.5,
  conf_level = 0.95,
  check.KM = TRUE,
  stop.onerror = FALSE,
  censoring_allmarks = TRUE
)
}
\description{
Creates counting process dataset
}
