% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_counting.R
\name{get_dfcounting}
\alias{get_dfcounting}
\title{Get df_counting}
\usage{
get_dfcounting(
  df,
  tte.name,
  event.name,
  treat.name,
  arms,
  by.risk = 12,
  cox.digits = 3,
  lr.digits = 3,
  qprob = 0.5,
  strata.name = NULL,
  weight.name = NULL,
  check.KM = TRUE
)
}
\description{
Get df_counting
}
