% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_calculations_helpers.R
\name{KM_estimates}
\alias{KM_estimates}
\title{Kaplan-Meier estimates and Greenwood variance}
\usage{
KM_estimates(ybar, nbar)
}
\arguments{
\item{ybar}{Number at risk at each time.}

\item{nbar}{Number of events at each time.}
}
\value{
List with survival and variance estimates.
}
\description{
Kaplan-Meier estimates and Greenwood variance
}
