% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplotting_helpers.R
\name{plot_km_curves_counting}
\alias{plot_km_curves_counting}
\title{KM plotting helper}
\usage{
plot_km_curves_counting(
  at.points,
  S0.KM,
  idx0,
  idv0,
  S1.KM,
  idx1,
  idv1,
  col.0,
  col.1,
  ltys,
  lwds,
  Xlab,
  Ylab,
  ylim,
  xlim,
  show.ticks = FALSE,
  cens0 = NULL,
  risk.points,
  risk.points.label,
  cens1 = NULL,
  S0.KM.full = NULL,
  S1.KM.full = NULL,
  censor.cex = 1,
  time.zero = 0,
  tpoints.add = c(0),
  ...
)
}
\description{
KM plotting helper
}
