% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_calculations_helpers.R
\name{wlr_dhat_estimates}
\alias{wlr_dhat_estimates}
\title{Weighted log-rank and KM difference at tzero}
\usage{
wlr_dhat_estimates(dfcounting, rho = 0, gamma = 0, tzero = 24)
}
\arguments{
\item{dfcounting}{Data frame with counting process columns.}

\item{rho}{Weighting parameter.}

\item{gamma}{Weighting parameter.}

\item{tzero}{Time point for difference.}
}
\value{
List with statistics, variances, covariance, and correlation.
}
\description{
Weighted log-rank and KM difference at tzero
}
