% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_calculations_helpers.R
\name{wlr_estimates}
\alias{wlr_estimates}
\title{Weighted log-rank estimates and variance}
\usage{
wlr_estimates(ybar0, ybar1, nbar0, nbar1, rho = 0, gamma = 0)
}
\arguments{
\item{ybar0}{Number at risk in group 0.}

\item{ybar1}{Number at risk in group 1.}

\item{nbar0}{Number of events in group 0.}

\item{nbar1}{Number of events in group 1.}

\item{rho}{Weighting parameter.}

\item{gamma}{Weighting parameter.}
}
\value{
List with log-rank statistic and variance.
}
\description{
Weighted log-rank estimates and variance
}
