% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km_calculations_helpers.R
\name{z_score_calculations}
\alias{z_score_calculations}
\title{Score test statistic for survival data}
\usage{
z_score_calculations(nbar0, ybar0, nbar1, ybar1)
}
\arguments{
\item{nbar0}{Number of events in group 0.}

\item{ybar0}{Number at risk in group 0.}

\item{nbar1}{Number of events in group 1.}

\item{ybar1}{Number at risk in group 1.}
}
\value{
List with z-score, score, and variance.
}
\description{
Score test statistic for survival data
}
